project/requirements.txt
streamlit
pypdf
requests
beautifulsoup4
sentence-transformers
numpy
pandas
chromadb
faiss-cpu
